update to chromium 83.0.4103.0

+++ chrome/browser/web_applications/web_app_icon_manager.cc
@@ -12,8 +12,8 @@
 #include "base/memory/scoped_refptr.h"
 #include "base/stl_util.h"
 #include "base/strings/stringprintf.h"
-#include "base/task/post_task.h"
 #include "base/task/task_traits.h"
+#include "base/task/thread_pool.h"
 #include "chrome/browser/web_applications/components/web_app_utils.h"
 #include "chrome/browser/web_applications/file_utils_wrapper.h"
 #include "chrome/browser/web_applications/web_app.h"
@@ -28,42 +28,37 @@ namespace web_app {
 
 namespace {
 
-constexpr base::FilePath::CharType kTempDirectoryName[] =
-    FILE_PATH_LITERAL("Temp");
-
-constexpr base::FilePath::CharType kIconsDirectoryName[] =
-    FILE_PATH_LITERAL("Icons");
-
-base::FilePath GetAppDirectory(const base::FilePath& web_apps_directory,
-                               const AppId& app_id) {
-  return web_apps_directory.AppendASCII(app_id);
-}
-
-base::FilePath GetTempDir(FileUtilsWrapper* utils,
-                          const base::FilePath& web_apps_dir) {
-  // Create the temp directory as a sub-directory of the WebApps directory.
-  // This guarantees it is on the same file system as the WebApp's eventual
-  // install target.
-  base::FilePath temp_path = web_apps_dir.Append(kTempDirectoryName);
-  if (utils->PathExists(temp_path)) {
-    if (!utils->DirectoryExists(temp_path)) {
-      LOG(ERROR) << "Not a directory: " << temp_path.value();
-      return base::FilePath();
+// Returns false if directory doesn't exist or it is not writable.
+bool CreateDirectoryIfNotExists(FileUtilsWrapper* utils,
+                                const base::FilePath& path) {
+  if (utils->PathExists(path)) {
+    if (!utils->DirectoryExists(path)) {
+      LOG(ERROR) << "Not a directory: " << path.value();
+      return false;
     }
-    if (!utils->PathIsWritable(temp_path)) {
-      LOG(ERROR) << "Can't write to path: " << temp_path.value();
-      return base::FilePath();
+    if (!utils->PathIsWritable(path)) {
+      LOG(ERROR) << "Can't write to path: " << path.value();
+      return false;
     }
     // This is a directory we can write to.
-    return temp_path;
+    return true;
   }
 
   // Directory doesn't exist, so create it.
-  if (!utils->CreateDirectory(temp_path)) {
-    LOG(ERROR) << "Could not create directory: " << temp_path.value();
-    return base::FilePath();
+  if (!utils->CreateDirectory(path)) {
+    LOG(ERROR) << "Could not create directory: " << path.value();
+    return false;
   }
-  return temp_path;
+  return true;
+}
+
+// This is a private implementation detail of WebAppIconManager, where and how
+// to store icon files.
+base::FilePath GetAppIconsDirectory(
+    const base::FilePath& app_manifest_resources_directory) {
+  constexpr base::FilePath::CharType kIconsDirectoryName[] =
+      FILE_PATH_LITERAL("Icons");
+  return app_manifest_resources_directory.Append(kIconsDirectoryName);
 }
 
 bool WriteIcon(FileUtilsWrapper* utils,
@@ -95,13 +90,14 @@ bool WriteIcon(FileUtilsWrapper* utils,
 bool WriteIcons(FileUtilsWrapper* utils,
                 const base::FilePath& app_dir,
                 const std::map<SquareSizePx, SkBitmap>& icon_bitmaps) {
-  const base::FilePath icons_dir = app_dir.Append(kIconsDirectoryName);
+  const base::FilePath icons_dir = GetAppIconsDirectory(app_dir);
   if (!utils->CreateDirectory(icons_dir)) {
     LOG(ERROR) << "Could not create icons directory.";
     return false;
   }
 
-  for (const std::pair<SquareSizePx, SkBitmap>& icon_bitmap : icon_bitmaps) {
+  for (const std::pair<const SquareSizePx, SkBitmap>& icon_bitmap :
+       icon_bitmaps) {
     if (!WriteIcon(utils, icons_dir, icon_bitmap.second))
       return false;
   }
@@ -115,10 +111,13 @@ bool WriteDataBlocking(const std::unique_ptr<FileUtilsWrapper>& utils,
                        const base::FilePath& web_apps_directory,
                        const AppId& app_id,
                        const std::map<SquareSizePx, SkBitmap>& icons) {
-  const base::FilePath temp_dir = GetTempDir(utils.get(), web_apps_directory);
-  if (temp_dir.empty()) {
-    LOG(ERROR)
-        << "Could not get path to WebApps temporary directory in profile.";
+  // Create the temp directory under the web apps root.
+  // This guarantees it is on the same file system as the WebApp's eventual
+  // install target.
+  base::FilePath temp_dir = GetWebAppsTempDirectory(web_apps_directory);
+  if (!CreateDirectoryIfNotExists(utils.get(), temp_dir)) {
+    LOG(ERROR) << "Could not create or write to WebApps temporary directory in "
+                  "profile.";
     return false;
   }
 
@@ -131,12 +130,20 @@ bool WriteDataBlocking(const std::unique_ptr<FileUtilsWrapper>& utils,
   if (!WriteIcons(utils.get(), app_temp_dir.GetPath(), icons))
     return false;
 
-  // Commit: move whole app data dir to final destination in one mv operation.
-  const base::FilePath app_dir = GetAppDirectory(web_apps_directory, app_id);
+  base::FilePath manifest_resources_directory =
+      GetManifestResourcesDirectory(web_apps_directory);
+  if (!CreateDirectoryIfNotExists(utils.get(), manifest_resources_directory)) {
+    LOG(ERROR) << "Could not create Manifest Resources directory.";
+    return false;
+  }
 
-  // Try to delete the destination. Needed for update.
+  base::FilePath app_dir =
+      GetManifestResourcesDirectoryForApp(web_apps_directory, app_id);
+
+  // Try to delete the destination. Needed for update. Ignore the result.
   utils->DeleteFileRecursively(app_dir);
 
+  // Commit: move whole app data dir to final destination in one mv operation.
   if (!utils->Move(app_temp_dir.GetPath(), app_dir)) {
     LOG(ERROR) << "Could not move temp WebApp directory to final destination.";
     return false;
@@ -151,15 +158,18 @@ bool WriteDataBlocking(const std::unique_ptr<FileUtilsWrapper>& utils,
 bool DeleteDataBlocking(const std::unique_ptr<FileUtilsWrapper>& utils,
                         const base::FilePath& web_apps_directory,
                         const AppId& app_id) {
-  const base::FilePath app_dir = GetAppDirectory(web_apps_directory, app_id);
+  base::FilePath app_dir =
+      GetManifestResourcesDirectoryForApp(web_apps_directory, app_id);
+
   return utils->DeleteFileRecursively(app_dir);
 }
 
 base::FilePath GetIconFileName(const base::FilePath& web_apps_directory,
                                const AppId& app_id,
                                int icon_size_px) {
-  const base::FilePath app_dir = GetAppDirectory(web_apps_directory, app_id);
-  const base::FilePath icons_dir = app_dir.Append(kIconsDirectoryName);
+  base::FilePath app_dir =
+      GetManifestResourcesDirectoryForApp(web_apps_directory, app_id);
+  base::FilePath icons_dir = GetAppIconsDirectory(app_dir);
 
   return icons_dir.AppendASCII(base::StringPrintf("%i.png", icon_size_px));
 }
@@ -259,7 +269,7 @@ std::vector<uint8_t> ReadCompressedIconBlocking(
 }
 
 constexpr base::TaskTraits kTaskTraits = {
-    base::ThreadPool(), base::MayBlock(), base::TaskPriority::USER_VISIBLE,
+    base::MayBlock(), base::TaskPriority::USER_VISIBLE,
     base::TaskShutdownBehavior::BLOCK_SHUTDOWN};
 
 }  // namespace
@@ -268,7 +278,7 @@ WebAppIconManager::WebAppIconManager(Profile* profile,
                                      WebAppRegistrar& registrar,
                                      std::unique_ptr<FileUtilsWrapper> utils)
     : registrar_(registrar), utils_(std::move(utils)) {
-  web_apps_directory_ = GetWebAppsDirectory(profile);
+  web_apps_directory_ = GetWebAppsRootDirectory(profile);
 }
 
 WebAppIconManager::~WebAppIconManager() = default;
@@ -278,7 +288,7 @@ void WebAppIconManager::WriteData(AppId app_id,
                                   WriteDataCallback callback) {
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
 
-  base::PostTaskAndReplyWithResult(
+  base::ThreadPool::PostTaskAndReplyWithResult(
       FROM_HERE, kTaskTraits,
       base::BindOnce(WriteDataBlocking, utils_->Clone(), web_apps_directory_,
                      std::move(app_id), std::move(icons)),
@@ -288,7 +298,7 @@ void WebAppIconManager::WriteData(AppId app_id,
 void WebAppIconManager::DeleteData(AppId app_id, WriteDataCallback callback) {
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
 
-  base::PostTaskAndReplyWithResult(
+  base::ThreadPool::PostTaskAndReplyWithResult(
       FROM_HERE, kTaskTraits,
       base::BindOnce(DeleteDataBlocking, utils_->Clone(), web_apps_directory_,
                      std::move(app_id)),
@@ -317,7 +327,7 @@ void WebAppIconManager::ReadIcons(
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
   DCHECK(HasIcons(app_id, icon_sizes_in_px));
 
-  base::PostTaskAndReplyWithResult(
+  base::ThreadPool::PostTaskAndReplyWithResult(
       FROM_HERE, kTaskTraits,
       base::BindOnce(ReadIconsBlocking, utils_->Clone(), web_apps_directory_,
                      app_id, icon_sizes_in_px),
@@ -333,7 +343,7 @@ void WebAppIconManager::ReadAllIcons(const AppId& app_id,
     return;
   }
 
-  base::PostTaskAndReplyWithResult(
+  base::ThreadPool::PostTaskAndReplyWithResult(
       FROM_HERE, kTaskTraits,
       base::BindOnce(ReadIconsBlocking, utils_->Clone(), web_apps_directory_,
                      app_id, web_app->downloaded_icon_sizes()),
@@ -349,7 +359,7 @@ void WebAppIconManager::ReadSmallestIcon(const AppId& app_id,
       FindDownloadedSizeInPxMatchBigger(app_id, icon_size_in_px);
   DCHECK(best_size_in_px.has_value());
 
-  base::PostTaskAndReplyWithResult(
+  base::ThreadPool::PostTaskAndReplyWithResult(
       FROM_HERE, kTaskTraits,
       base::BindOnce(ReadIconBlocking, utils_->Clone(), web_apps_directory_,
                      app_id, best_size_in_px.value()),
@@ -366,7 +376,7 @@ void WebAppIconManager::ReadSmallestCompressedIcon(
       FindDownloadedSizeInPxMatchBigger(app_id, icon_size_in_px);
   DCHECK(best_size_in_px.has_value());
 
-  base::PostTaskAndReplyWithResult(
+  base::ThreadPool::PostTaskAndReplyWithResult(
       FROM_HERE, kTaskTraits,
       base::BindOnce(ReadCompressedIconBlocking, utils_->Clone(),
                      web_apps_directory_, app_id, best_size_in_px.value()),
@@ -395,7 +405,7 @@ void WebAppIconManager::ReadIconAndResize(const AppId& app_id,
 
   DCHECK(best_downloaded_size.has_value());
 
-  base::PostTaskAndReplyWithResult(
+  base::ThreadPool::PostTaskAndReplyWithResult(
       FROM_HERE, kTaskTraits,
       base::BindOnce(ReadIconAndResizeBlocking, utils_->Clone(),
                      web_apps_directory_, app_id, best_downloaded_size.value(),
