+++ gpu/command_buffer/service/skia_utils.cc
@@ -20,6 +20,7 @@
 #include "gpu/vulkan/vulkan_device_queue.h"
 #include "gpu/vulkan/vulkan_fence_helper.h"
 #include "gpu/vulkan/vulkan_function_pointers.h"
+#include "gpu/vulkan/vulkan_image.h"
 #endif
 
 namespace gpu {
@@ -42,6 +43,11 @@ void CleanupAfterSkiaFlush(void* context) {
 template <class T>
 void DeleteSkObject(SharedContextState* context_state, sk_sp<T> sk_object) {
   DCHECK(sk_object && sk_object->unique());
+
+  if (context_state->context_lost())
+    return;
+  DCHECK(!context_state->gr_context()->abandoned());
+
   if (!context_state->GrContextIsVulkan())
     return;
 
@@ -128,6 +134,11 @@ void AddVulkanCleanupTaskForSkiaFlush(
 void DeleteGrBackendTexture(SharedContextState* context_state,
                             GrBackendTexture* backend_texture) {
   DCHECK(backend_texture && backend_texture->isValid());
+
+  if (context_state->context_lost())
+    return;
+  DCHECK(!context_state->gr_context()->abandoned());
+
   if (!context_state->GrContextIsVulkan()) {
     context_state->gr_context()->deleteBackendTexture(
         std::move(*backend_texture));
@@ -140,9 +151,8 @@ void DeleteGrBackendTexture(SharedContextState* context_state,
   fence_helper->EnqueueCleanupTaskForSubmittedWork(base::BindOnce(
       [](const sk_sp<GrContext>& gr_context, GrBackendTexture backend_texture,
          gpu::VulkanDeviceQueue* device_queue, bool is_lost) {
-        // If underlying Vulkan device is destroyed, gr_context should have been
-        // abandoned, the deleteBackendTexture() should be noop.
-        gr_context->deleteBackendTexture(std::move(backend_texture));
+        if (!gr_context->abandoned())
+          gr_context->deleteBackendTexture(std::move(backend_texture));
       },
       sk_ref_sp(context_state->gr_context()), std::move(*backend_texture)));
 #endif
@@ -158,6 +168,20 @@ void DeleteSkSurface(SharedContextState* context_state,
 }
 
 #if BUILDFLAG(ENABLE_VULKAN)
+GrVkImageInfo CreateGrVkImageInfo(VulkanImage* image) {
+  DCHECK(image);
+  VkPhysicalDevice physical_device =
+      image->device_queue()->GetVulkanPhysicalDevice();
+  GrVkYcbcrConversionInfo gr_ycbcr_info = CreateGrVkYcbcrConversionInfo(
+      physical_device, image->image_tiling(), image->ycbcr_info());
+  GrVkAlloc alloc(image->device_memory(), /*offset=*/0, image->device_size(),
+                  /*flags=*/0);
+  bool is_protected = image->flags() & VK_IMAGE_CREATE_PROTECTED_BIT;
+  return GrVkImageInfo(
+      image->image(), alloc, image->image_tiling(), image->image_layout(),
+      image->format(), /*levelCount=*/1, image->queue_family_index(),
+      is_protected ? GrProtected::kYes : GrProtected::kNo, gr_ycbcr_info);
+}
 
 GrVkYcbcrConversionInfo CreateGrVkYcbcrConversionInfo(
     VkPhysicalDevice physical_device,
